mostly finished image generation part

+++ other_scripts/long_tour_process.py
@@ -1,26 +1,32 @@
-from enum import unique
 from itertools import pairwise
 import shutil
 import argparse
-import math
 from PIL import Image
 import imagehash
-import alive_progress
 import os 
 from collections import Counter
 import multiprocessing
 from joblib import Parallel, delayed
 from tqdm import tqdm
 from bisect import bisect_left
-
+from pathlib import Path
+import res.file_manip_resources as rfm
+import re
 
 parser = argparse.ArgumentParser(description="useful image tools")
+parser.add_argument('--name', help="name of output directory", type=str)
 parser.add_argument('--dir', default=r'.\image_folder\in', help="input directiory", type=str)
 parser.add_argument('--outdir', default=r'.\image_folder\out', help="output directory", type=str)
 parser.add_argument('--delimiter', default=r'.\image_folder\delimiter\del.png', help="output directory", type=str)
 parser.add_argument('--verbose', action=argparse.BooleanOptionalAction)
 opt = parser.parse_args()
 
+#set default value for --name
+if not opt.name:
+    opt.name = "out"
+
+#magic values
+#change at own risk
 threshold = 15
 delimiter_threshold = 10
 odd_threshold = 5
@@ -77,39 +83,50 @@ def take_closest(myList, myNumber):
     else:
         return before
 
+opt.outdir = rfm.unique_name_generate(opt.outdir, opt.name)
+#ensure paths exist
+Path(opt.outdir).mkdir(parents=True, exist_ok=True)
+Path(opt.dir).mkdir(parents=True, exist_ok=True)
+Path(os.path.dirname(opt.delimiter)).mkdir(parents=True, exist_ok=True)
+
+#check delimiter exists
+if (not os.path.isfile(opt.delimiter)):
+    raise Exception("\n\nFailed to find delimiter.\n")
+
+#check input directory has files in it
+if os.path.isdir(opt.dir) and os.path.exists(opt.dir):
+    if len(os.listdir(opt.dir)) == 0:
+        raise Exception("\n\nNo files in input directory\n")
+
 #delete files in out directory
+#sanity check, should normally be empty
 delete_directory_content(opt.outdir)
 
-#generate image hash of each image
-f = get_files(opt.dir)
-image_hashes = []
+#Generate hash of delimiter
 try:
     delimiter = imagehash.phash(Image.open(opt.delimiter))
 except:
-    raise Exception("Failed to load and hash delimiter.")
+    raise Exception("\n\nFailed to load and hash delimiter.\n")
 
+#get files in input directory
+f = get_files(opt.dir)
+image_hashes = []
 paths = [direntry.path for direntry in f]
 
+#calculate hashes for each file in input directory
 def phash_calc(file_obj):
-    h = imagehash.phash(Image.open(file_obj))
-    return h
+    return imagehash.phash(Image.open(file_obj))
 
 num_cores = multiprocessing.cpu_count()
-
 image_hashes = Parallel(n_jobs=num_cores)(delayed(phash_calc)(i) for i in tqdm(paths, desc="Hashing Images: "))
 
+#generate dictionary of hash to index
 image_hashes_dict = {}
 for idx, hsh in enumerate(image_hashes):
     image_hashes_dict[str(hsh)] = idx
 
-#with alive_progress.alive_bar(len(f), title="Generating Hashes", length=100) as bar:
-#    for idx, n in enumerate(f):
-#        h = hash_func(Image.open(n.path))
-#        image_hashes.append(h)
-#        bar()
-
+#Removes images that are not similar to any other image
 counter = Counter(image_hashes)
-
 hashes_filtered = image_hashes.copy()
 unique_hashes = []
 unique_hashes_set = set()
@@ -130,31 +147,18 @@ for key in tqdm(reversed(counter.keys()), desc="Removing Loners: ", total=len(co
                 has_close = True
 
         if not has_close: #skip removal
-            #print(f[idx].name)
             hashes_filtered.remove(key)
 
+#removes images that are exactly the same as other images
+if opt.verbose: print("Removing identicle images")
 for v in tqdm(hashes_filtered, desc="Removing Duplicates: "): #remove duplicate values
     if v not in unique_hashes_set or abs(v - delimiter) <= delimiter_threshold:
         unique_hashes.append(v)
         unique_hashes_set.add(v)
 
-#print(len(unique_hashes))
-
-#for key in counter_removal:
-#    del counter[key]
-
-#for v in unique_hashes:
-    #print(str(v))
-
-if opt.verbose: print("Removing identicle images")
-
-#unique_hashes = list(dict.fromkeys(hashes_filtered))
-#unique_hashes = list(counter.keys())
-
+#remove images that are similar and positionally close to each other
 if opt.verbose: print("Evaluating neighbours")
 difference_arr = [abs(y-x) for (x, y) in pairwise(unique_hashes)]
-
-#print("Uniques: ", len(unique_hashes))
 
 #remove similar 
 if opt.verbose: print("Removing similar neighbours")
@@ -162,15 +166,9 @@ removed_count = 0
 for i in reversed(range(0, len(unique_hashes)-1)):
      if difference_arr[i] <= threshold:
         removed_count += 1
-        #if (abs(unique_hashes[i]-delimiter) <= delimiter_threshold):
-           #idx = image_hashes_dict[str(unique_hashes[i])]
-           #print(f[idx].name, difference_arr[i], i)
-        #else:
-        #    unique_hashes.pop(i)
         unique_hashes.pop(i)
 
-#print("Removed: ", removed_count)
-
+#seperate images by the delimiter image
 if opt.verbose: print("Seperating by delimiter")
 subdivide_result = []
 subdivide_temp_arr = []
@@ -182,30 +180,21 @@ for element in subdivide_itr:
 
         subdivide_result.append(subdivide_temp_arr)
         subdivide_temp_arr = []
-        #next(subdivide_itr, None)
         continue
     else:
         subdivide_temp_arr.append(element)
-#subdivide_result.append(unique_hashes)
 
 def collect_images(hash_val, unfiltered_hash_dict, paths_arr):
     i = unfiltered_hash_dict[str(hash_val)]
     return paths_arr[i]
 
 def save_images(filename, path):
-    
     shutil.copy(filename, path)
 
 if opt.verbose: print("Saving Files")
-#for idx, n in enumerate(subdivide_result):
 
-
-
-#for i in subdivide_result:
-#    print(len(i))
-#print(len(subdivide_result))
-#Parallel(n_jobs=num_cores)(delayed(save_result)(i) for i in indexed_subdivide_result)
-
+#save images
+#this can be parallelized however it doesn't improve speed
 for idx, hash_arr in enumerate(subdivide_result):
     path = opt.outdir + "\\" + str(idx)
     print(path, len(hash_arr))
@@ -220,11 +209,3 @@ for idx, hash_arr in enumerate(subdivide_result):
 
     for p in out_paths:
         save_images(p, path)
-
-    #out_paths = Parallel(n_jobs=num_cores)(delayed(collect_images)(h, image_hashes, paths) for h in tqdm(file[1]))
-    #print(len(out_paths))
-    #with ThreadPoolExecutor(10) as exe:
-            # submit all copy tasks
-            #_ = [exe.submit(save_images, filename, path) for filename in tqdm(out_paths)]
-
-    #Parallel(n_jobs=num_cores)(delayed(save_images)(p, path) for p in tqdm(out_paths))
